// packages/core/src/schemas/shared.schema.ts
import { z } from "zod";

export const commitSchema = z.object({
	id: z.string(),
	message: z.string(),
	treeId: z.string(),
	distinct: z.boolean(),
	url: z.string(),
});

export type CommitData = z.infer<typeof commitSchema>;

export const branchRefSchema = z.object({
	label: z.string(),
	ref: z.string(),
	sha: z.string(),
});

export type BranchRefData = z.infer<typeof branchRefSchema>;

export const pullRequestSchema = z.object({
	number: z.number(),
	state: z.string(),
	title: z.string(),
	head: branchRefSchema,
	base: branchRefSchema,
});

export type PullRequestData = z.infer<typeof pullRequestSchema>;

export const repositorySchema = z.object({
	name: z.string(),
	full_name: z.string(),
	owner: z.object({ login: z.string() }),
});

export type RepositoryData = z.infer<typeof repositorySchema>;